--- src/test/java/wooteco/subway/member/CustomMemberAcceptanceTest.java
@@ -1,152 +1,0 @@
-package wooteco.subway.member;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static wooteco.subway.auth.AuthAcceptanceTest.로그인되어_있음;
-
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import wooteco.subway.AcceptanceTest;
-import wooteco.subway.auth.dto.TokenResponse;
-import wooteco.subway.member.dto.MemberRequest;
-import wooteco.subway.member.dto.MemberResponse;
-
-public class CustomMemberAcceptanceTest extends AcceptanceTest {
-
-    public static final String EMAIL = "email@email.com";
-    public static final String PASSWORD = "password";
-    public static final int AGE = 20;
-    public static final String NEW_EMAIL = "new_email@email.com";
-    public static final String NEW_PASSWORD = "new_password";
-    public static final int NEW_AGE = 30;
-
-    @DisplayName("회원 정보를 관리한다.")
-    @Test
-    void manageMember() {
-        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE);
-        회원_생성됨(createResponse);
-
-        TokenResponse 사용자 = 로그인되어_있음(EMAIL, PASSWORD);
-
-        ExtractableResponse<Response> findResponse = 내_회원_정보_조회_요청(사용자);
-        회원_정보_조회됨(findResponse, EMAIL, AGE);
-
-        ExtractableResponse<Response> updateResponse = 내_회원_정보_수정_요청(사용자, EMAIL, NEW_PASSWORD, NEW_AGE);
-        회원_정보_수정됨(updateResponse);
-
-        ExtractableResponse<Response> deleteResponse = 내_회원_삭제_요청(사용자);
-        회원_삭제됨(deleteResponse);
-    }
-
-    @DisplayName("회원 가입 시도시 이메일 정보가 null 이거나 유효한 형식이 아닌 경우")
-    @Test
-    void validateJoinMemberEmail() {
-        assertThat(회원_생성을_요청(null, PASSWORD, AGE).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(회원_생성을_요청("", PASSWORD, AGE).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(회원_생성을_요청(" ", PASSWORD, AGE).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(회원_생성을_요청("mail.com", PASSWORD, AGE).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("회원 가입 시도시 비밀번호 정보가 null 이거나 공백인 경우")
-    @Test
-    void validateJoinMemberPassword() {
-        assertThat(회원_생성을_요청(EMAIL, null, AGE).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(회원_생성을_요청(EMAIL, "", AGE).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(회원_생성을_요청(EMAIL, " ", AGE).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("회원 가입 시도시 나이 정보가 null 이거나 음수인 경우")
-    @Test
-    void validateJoinMemberAge() {
-        assertThat(회원_생성을_요청(EMAIL, PASSWORD, null).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(회원_생성을_요청(EMAIL, PASSWORD, -1).statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("회원 가입 시도시 중복된 이메일이 이미 존재하는지 확인")
-    @Test
-    void validateDuplicateEmail() {
-        assertThat(회원_생성을_요청(EMAIL, PASSWORD, AGE).statusCode()).isEqualTo(HttpStatus.CREATED.value());
-
-        RestAssured
-            .given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/api/members/check-validation?email=" + "dododo@naver.com")
-            .then().log().all()
-            .statusCode(HttpStatus.OK.value());
-
-        RestAssured
-            .given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/api/members/check-validation?email=" + EMAIL)
-            .then().log().all()
-            .statusCode(HttpStatus.BAD_REQUEST.value());
-    }
-
-    public static ExtractableResponse<Response> 회원_생성을_요청(String email, String password, Integer age) {
-        MemberRequest memberRequest = new MemberRequest(email, password, age);
-
-        return RestAssured
-            .given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(memberRequest)
-            .when().post("/api/members")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> 내_회원_정보_조회_요청(TokenResponse tokenResponse) {
-        return RestAssured
-            .given().log().all()
-            .auth().oauth2(tokenResponse.getAccessToken())
-            .accept(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/api/members/me")
-            .then().log().all()
-            .statusCode(HttpStatus.OK.value())
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> 내_회원_정보_수정_요청(TokenResponse tokenResponse, String email, String password, Integer age) {
-        MemberRequest memberRequest = new MemberRequest(email, password, age);
-
-        return RestAssured
-            .given().log().all()
-            .auth().oauth2(tokenResponse.getAccessToken())
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(memberRequest)
-            .when().put("/api/members/me")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> 내_회원_삭제_요청(TokenResponse tokenResponse) {
-        return RestAssured
-            .given().log().all()
-            .auth().oauth2(tokenResponse.getAccessToken())
-            .when().delete("/api/members/me")
-            .then().log().all()
-            .extract();
-    }
-
-    public static void 회원_생성됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    public static void 회원_정보_조회됨(ExtractableResponse<Response> response, String email, int age) {
-        MemberResponse memberResponse = response.as(MemberResponse.class);
-        assertThat(memberResponse.getId()).isNotNull();
-        assertThat(memberResponse.getEmail()).isEqualTo(email);
-        assertThat(memberResponse.getAge()).isEqualTo(age);
-    }
-
-    public static void 회원_정보_수정됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 회원_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-}
